requirements.txt
python-telegram-bot
discord.py
python-dotenv
nest-asyncio
